requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
python-dotenv>=0.19.0